requirements.txt
pandas 
numpy 
pillow 
sentence-transformers